requirements.txt
# Core Dependencies
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.79

# PDF Processing
pypdf==3.17.4

# Vector Store
faiss-cpu==1.7.4

# Local Embeddings (No API costs!)
sentence-transformers==2.7.0
huggingface-hub==0.23.0

# Anthropic Claude Integration
anthropic==0.39.0
langchain-anthropic==0.3.0

# Web UI
streamlit==1.29.0

# Utilities
python-dotenv==1.0.0
